// src/enums/sidebar.enum.ts
export enum SidebarItemTypeEnum {
  LINK = 'LINK',
  MENU = 'MENU',
  GROUP = 'GROUP',
  HEADER = 'HEADER',
  DIVIDER = 'DIVIDER',
}

export enum SidebarVisibilityEnum {
  VISIBLE = 'VISIBLE',
  DISABLED = 'DISABLED',
  HIDDEN = 'HIDDEN',
  SYSTEM = 'SYSTEM',
}

export enum SidebarPermissionOperatorEnum {
  OR = 'OR',
  AND = 'AND',
}

export enum SidebarBadgeTypeEnum {
  INFO = 'INFO',
  NEW = 'NEW',
  BETA = 'BETA',
  ALPHA = 'ALPHA',
  COUNT = 'COUNT',
}